rest_lab6_python/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0
pydantic[email]==1.10.13
